added files via upload
named my files incorrectly (scripts.js and styles.css), so wasn't able to commit after I realized my mistake and renamed them at the end of the task when I realized that was the reason the app was not working. Unfortunately, this means that there are no commits for this project.